Add index.js to every component. Refactor imports to not have double naming.
