+++ src/components/App/index.js
@@ -1,3 +1,1 @@
-import App from './App';
-
-export default App;
+export { default } from './App';
